% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLongRunVar.R
\name{estimateLongRunVar}
\alias{estimateLongRunVar}
\title{Estimate long-run variance}
\usage{
estimateLongRunVar(x, bw = NULL)
}
\description{
Routine for estimating the long-run variance.
}
\keyword{internal}
