% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRho.R
\name{estimateRho}
\alias{estimateRho}
\title{Estimate \code{rho} coefficient}
\usage{
estimateRho(Phi, rhoLimit)
}
\arguments{
\item{Phi}{Matrix of computed contrasts generated by \code{tsACV()}.}

\item{rhoLimit}{Parameter \code{rhoLimit} limits to the absolute value of the estimated \code{rho} coefficient. This is useful as estimated values very close to 1 might cause instability.}
}
\value{
Estimated \code{rho} coefficient (numeric vector of length 1).
}
\description{
Internal function for estimating the rho coefficient.
}
\keyword{internal}
