% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoPhi.R
\name{infoPhi}
\alias{infoPhi}
\title{Recover information about \code{Phi}}
\usage{
infoPhi(Phi)
}
\description{
Recovers all the necessary parameters using which the \code{Phi} was constructed.
}
\keyword{internal}
