% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.estimateL.R
\name{print.estimateL}
\alias{print.estimateL}
\title{Printing method for \code{estimateL}}
\usage{
\method{print}{estimateL}(x, ...)
}
\description{
Prettify the output of \code{estimateL} when printed.
}
\keyword{internal}
