% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.testL.R
\name{print.testL}
\alias{print.testL}
\title{Printing method for class \code{"testL"}}
\usage{
\method{print}{testL}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"testL"}.}
}
\value{
Does not return a value. It is used to print out the test results.
}
\description{
Internal printing method for \code{"testL"} object generated by \code{testL()}.
}
\keyword{internal}
