% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftMatrix.R
\name{shiftMatrix}
\alias{shiftMatrix}
\title{Construct shift matrix}
\usage{
shiftMatrix(n, q)
}
\arguments{
\item{n}{Integer specifying dimensions of the shift matrix.}

\item{q}{Integer specifying the order of the shift matrix. Value \code{q = 1} (resp. \code{q = -1}) indicates the upper (resp. lower) shift matrix. Larger (resp. smaller) values represent higher powers of the respective shift matrices.}
}
\value{
Returns a sparse matrix (class \code{"ngCMatrix"}).
}
\description{
Internal function for creation of sparse shift matrix.
}
\keyword{internal}
