% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateLongRunVar.R
\name{estimateLongRunVar}
\alias{estimateLongRunVar}
\title{Estimate long-run variance}
\usage{
estimateLongRunVar(x, bw = NULL)
}
\arguments{
\item{x}{Univariate time-series object.}

\item{bw}{Bandwidth for long run variance estimation.}
}
\value{
Estimated long run variance (numeric vector of length 1).
}
\description{
Internal function for estimating the long-run variance.
}
\keyword{internal}
