% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoPhi.R
\name{infoPhi}
\alias{infoPhi}
\title{Recover information about \code{Phi}}
\usage{
infoPhi(Phi)
}
\arguments{
\item{Phi}{Matrix of computed contrasts generated by \code{tsACV()}.}
}
\value{
List of parameters that were used to generate \code{Phi}.
}
\description{
Internal function which recovers all the necessary parameters using which the \code{Phi} was constructed and some additional useful variables derived from these parameters.
}
\keyword{internal}
