% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.estimateL.R
\name{print.estimateL}
\alias{print.estimateL}
\title{Printing method for class \code{"estimateL"}}
\usage{
\method{print}{estimateL}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"estimateL"}.}
}
\value{
Does not return a value. It is used to print out the loss estimate along its standard error and confidence interval.
}
\description{
Internal printing method for \code{"estimateL"} object generated by \code{estimateL()}.
}
\keyword{internal}
